1.fix(cnn): not compatible with azure python environment
