+++ hw/cnn/cifar.yml
@@ -2,9 +2,7 @@ $schema: https://azuremlschemas.azureedge.net/latest/commandJob.schema.json
 code:
   local_path: ./
 command: >-
-  python start.py
-  --epochs 128
-  --data_dir {inputs.cifar}
+  python start.py --epochs 128 --data_dir {inputs.cifar};
 inputs:
   cifar:
     data:
